Add CMake setting for iOS product bundle identifiers
The iOS product bundle identifiers need to be changed when building locally for development purposes but are hard-coded as "com.BabylonNative.Playground.iOS" and "com.BabylonNative.ValidationTests.iOS".

This change adds CMake settings for the iOS product bundle identifiers so they can be changed without modifying the CMake files directly.